# jclouds-sphereon-storage/swagger.json
{
  "basePath": "/bucket-storage/0.8",
  "paths": {
    "/containers/{containerId}": {
      "get": {
        "summary": "Get container information",
        "security": [{"oauth2schema": ["global"]}],
        "produces": ["application/json;charset=UTF-8"],
        "description": "Get information on a container",
        "operationId": "getContainer",
        "responses": {
          "200": {
            "schema": {"$ref": "#/definitions/ContainerResponse"},
            "description": "Container info returned"
          },
          "400": {
            "schema": {"$ref": "#/definitions/ErrorResponse"},
            "description": "Invalid container name or id"
          },
          "404": {
            "schema": {"$ref": "#/definitions/ErrorResponse"},
            "description": "Container does not exist"
          }
        },
        "parameters": [{
          "in": "path",
          "name": "containerId",
          "description": "containerId",
          "type": "string",
          "required": true
        }],
        "tags": ["Storage"],
        "consumes": ["application/json"]
      },
      "delete": {
        "summary": "Delete an existing container",
        "security": [{"oauth2schema": ["global"]}],
        "produces": ["application/json;charset=UTF-8"],
        "description": "Delete an existing container including the stored streams if indicated.",
        "operationId": "deleteContainer",
        "responses": {
          "200": {
            "schema": {"$ref": "#/definitions/ContainerResponse"},
            "description": "Container deleted"
          },
          "400": {
            "schema": {"$ref": "#/definitions/ErrorResponse"},
            "description": "Invalid container name or container was not empty and recursion was not set"
          },
          "404": {
            "schema": {"$ref": "#/definitions/ErrorResponse"},
            "description": "Container was not found"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "containerId",
            "description": "containerId",
            "type": "string",
            "required": true
          },
          {
            "in": "query",
            "name": "delete",
            "description": "delete",
            "type": "string",
            "required": false,
            "enum": [
              "RECURSIVE",
              "EMPTY_ONLY"
            ]
          }
        ],
        "tags": ["Storage"],
        "consumes": ["application/json"]
      },
      "put": {
        "summary": "Update a container",
        "security": [{"oauth2schema": ["global"]}],
        "produces": ["application/json;charset=UTF-8"],
        "description": "Update an existing container with new container settings",
        "operationId": "updateContainer",
        "responses": {
          "200": {
            "schema": {"$ref": "#/definitions/ContainerResponse"},
            "description": "Container updated"
          },
          "400": {
            "schema": {"$ref": "#/definitions/ErrorResponse"},
            "description": "Invalid Container request"
          },
          "404": {
            "schema": {"$ref": "#/definitions/ErrorResponse"},
            "description": "Backend was not found"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "containerId",
            "description": "containerId",
            "type": "string",
            "required": true
          },
          {
            "schema": {"$ref": "#/definitions/ContainerRequest"},
            "in": "body",
            "name": "containerRequest",
            "description": "containerRequest",
            "required": true
          }
        ],
        "tags": ["Storage"],
        "consumes": ["application/json;charset=UTF-8"]
      }
    },
    "/containers": {"post": {
      "summary": "Create a new container",
      "security": [{"oauth2schema": ["global"]}],
      "produces": ["application/json;charset=UTF-8"],
      "description": "Create a new container",
      "operationId": "createContainer",
      "responses": {
        "201": {
          "schema": {"$ref": "#/definitions/ContainerResponse"},
          "description": "Container created"
        },
        "400": {
          "schema": {"$ref": "#/definitions/ErrorResponse"},
          "description": "Invalid Container request"
        },
        "404": {
          "schema": {"$ref": "#/definitions/ErrorResponse"},
          "description": "Backend was not found"
        }
      },
      "parameters": [{
        "schema": {"$ref": "#/definitions/ContainerRequest"},
        "in": "body",
        "name": "containerRequest",
        "description": "containerRequest",
        "required": true
      }],
      "tags": ["Storage"],
      "consumes": ["application/json;charset=UTF-8"]
    }},
    "/info/path/{containerId}/": {"get": {
      "summary": "List streams in root",
      "security": [{"oauth2schema": ["global"]}],
      "produces": ["application/json"],
      "description": "Returns strean information from the root. When the path points to a file (and if the file exists,) a StreamInfo record of that single file will be returned. When the path points to a folder, the StreamInfo records of all files in that folder will be returned.",
      "operationId": "listInfoUsingRoot",
      "responses": {
        "200": {
          "schema": {"$ref": "#/definitions/InfoResponse"},
          "description": "Object information result"
        },
        "400": {
          "schema": {"$ref": "#/definitions/ErrorResponse"},
          "description": "Invalid container or stream information request"
        },
        "404": {
          "schema": {"$ref": "#/definitions/ErrorResponse"},
          "description": "Container was not found"
        }
      },
      "parameters": [{
        "in": "path",
        "name": "containerId",
        "description": "containerId",
        "type": "string",
        "required": true
      }],
      "tags": ["Storage"],
      "consumes": ["application/json"]
    }},
...
    "/streams/path/{containerId}/{path*}": {
      "post": {
        "summary": "Create a new stream within a container",
        "security": [{"oauth2schema": ["global"]}],
        "produces": ["application/json"],
        "description": "Create a new stream within a container. Query parameter creationMode accepts values NEW_ONLY to create a new stream, REQUIRE_EXISTING to update an stream and ALLOW_EXISTING to create a new stream or overwrite an existing stream.",
        "operationId": "createStreamUsingPath",
        "responses": {
          "201": {
            "schema": {"$ref": "#/definitions/StreamResponse"},
            "description": "Object created"
          },
          "400": {
            "schema": {"$ref": "#/definitions/ErrorResponse"},
            "description": "Invalid container or stream request"
          },
          "404": {
            "schema": {"$ref": "#/definitions/ErrorResponse"},
            "description": "Container was not found"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "containerId",
            "description": "containerId",
            "type": "string",
            "required": true
          },
          {
            "in": "path",
            "name": "path*",
            "description": "path*",
            "type": "string",
            "required": true
          },
          {
            "in": "query",
            "name": "creationMode",
            "description": "creationMode",
            "type": "string",
            "required": false,
            "enum": [
              "NEW_ONLY",
              "REQUIRE_EXISTING",
              "ALLOW_EXISTING"
            ]
          },
          {
            "in": "formData",
            "name": "stream",
            "description": "stream",
            "type": "file",
            "required": true
          }
        ],
        "tags": ["Storage"],
        "consumes": ["multipart/form-data"]
      },
      "get": {
        "summary": "Get an stream from a container",
        "security": [{"oauth2schema": ["global"]}],
        "produces": [
          "application/octet-stream",
          "application/json;charset=UTF-8"
        ],
        "description": "Get an existing stream from a container",
        "operationId": "getStreamUsingPath",
        "responses": {
          "200": {
            "schema": {
              "format": "byte",
              "type": "string"
            },
            "description": "Object retrieved"
          },
          "400": {
            "schema": {"$ref": "#/definitions/ErrorResponse"},
            "description": "Invalid container or stream name"
          },
          "404": {
            "schema": {"$ref": "#/definitions/ErrorResponse"},
            "description": "Object or container did not exist"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "containerId",
            "description": "containerId",
            "type": "string",
            "required": true
          },
          {
            "in": "path",
            "name": "path*",
            "description": "path*",
            "type": "string",
            "required": true
          }
        ],
        "tags": ["Storage"],
        "consumes": ["application/json"]
      },
      "delete": {
        "summary": "Delete an existing stream from a container.",
        "security": [{"oauth2schema": ["global"]}],
        "produces": ["*/*"],
        "description": "Delete an existing stream from a container. There is no request nor response body for a delete request.",
        "operationId": "deleteStreamUsingPath",
        "responses": {
          "200": {"description": "Object deleted"},
          "400": {
            "schema": {"$ref": "#/definitions/ErrorResponse"},
            "description": "Invalid container or stream name"
          },
          "404": {
            "schema": {"$ref": "#/definitions/ErrorResponse"},
            "description": "Container or stream was not found"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "containerId",
            "description": "containerId",
            "type": "string",
            "required": true
          },
          {
            "in": "path",
            "name": "path*",
            "description": "path*",
            "type": "string",
            "required": true
          }
        ],
        "tags": ["Storage"],
        "consumes": ["application/json"]
      }
    },
...
  },
  "host": "gw.api.cloud.sphereon.com",
  "schemes": ["https"],
  "securityDefinitions": {"oauth2schema": {
    "tokenUrl": "https://gw.api.cloud.sphereon.com/token",
    "scopes": {"global": "accessEverything"},
    "type": "oauth2",
    "flow": "application"
  }},
  "definitions": {
...
    "InfoResponse": {
      "type": "object",
      "properties": {"streamData": {
        "type": "array",
        "items": {"$ref": "#/definitions/StreamData"}
      }}
    },
    "ContainerResponse": {
      "type": "object",
      "properties": {
        "container": {"$ref": "#/definitions/Container"},
        "state": {
          "type": "string",
          "enum": [
            "REQUEST",
            "CREATED",
            "UPDATED",
            "DELETED"
          ]
        }
      }
    },
...
    "StreamResponse": {
      "type": "object",
      "properties": {
        "streamLocation": {"$ref": "#/definitions/StreamLocation"},
        "contentLength": {
          "format": "int64",
          "type": "integer"
        },
        "state": {
          "type": "string",
          "enum": [
            "REQUEST",
            "CREATED",
            "UPDATED",
            "DELETED"
          ]
        },
        "contentType": {"type": "string"}
      }
    },
...
    "Error": {
      "description": "An error",
      "type": "object",
      "required": [
        "code",
        "level",
        "message"
      ],
      "properties": {
        "code": {"type": "string"},
        "level": {
          "type": "string",
          "enum": [
            "INFO",
            "WARNING",
            "FATAL"
          ]
        },
        "cause": {"$ref": "#/definitions/Error"},
        "message": {"type": "string"}
      }
    },
...
    "StreamLocation": {
      "description": "Location record of data stream",
      "type": "object",
      "properties": {
        "folderPath": {"type": "string"},
        "originalFileName": {"type": "string"},
        "fileName": {"type": "string"},
        "id": {
          "readOnly": true,
          "type": "string"
        },
        "containerId": {"type": "string"}
      }
    },
    "ErrorResponse": {
      "description": "The error response",
      "type": "object",
      "properties": {"errors": {
        "type": "array",
        "items": {"$ref": "#/definitions/Error"}
      }}
    },
...
    "Container": {
      "type": "object",
      "properties": {
        "name": {"type": "string"},
        "backendId": {
          "description": "The Sphereon backend in which the container will be stored or is stored.",
          "type": "string"
        },
        "id": {"type": "string"}
      }
    },
    "StreamData": {
      "description": "Information record from a data stream",
      "type": "object",
      "properties": {
        "streamLocation": {"$ref": "#/definitions/StreamLocation"},
        "containerName": {"type": "string"},
        "providerId": {"type": "string"},
        "streamLength": {
          "format": "int64",
          "type": "integer"
        },
        "etag": {"type": "string"},
        "timeCreated": {
          "format": "date-time",
          "type": "string"
        },
        "region": {
          "type": "string",
          "enum": [
            "AMAZON_S3_AP_NORTHEAST_1",
            "AMAZON_S3_AP_NORTHEAST_2",
            "AMAZON_S3_AP_SOUTH_1",
            "AMAZON_S3_AP_SOUTHEAST_1",
            "AMAZON_S3_AP_SOUTHEAST_2",
            "AMAZON_S3_CA_CENTRAL_1",
            "AMAZON_S3_CN_NORTH_1",
            "AMAZON_S3_EU_CENTRAL_1",
            "AMAZON_S3_EU_WEST_1",
            "AMAZON_S3_EU_WEST_2",
            "AMAZON_S3_SA_EAST_1",
            "AMAZON_S3_US_EAST_1",
            "AMAZON_S3_US_EAST_2",
            "AMAZON_S3_US_STANDARD",
            "AMAZON_S3_US_WEST_1",
            "AMAZON_S3_US_WEST_2",
            "GOOGLE_STORAGE_ASIA",
            "GOOGLE_STORAGE_ASIA_EAST1",
            "GOOGLE_STORAGE_EU",
            "GOOGLE_STORAGE_US",
            "GOOGLE_STORAGE_US_CENTRAL1",
            "GOOGLE_STORAGE_US_CENTRAL2",
            "GOOGLE_STORAGE_US_EAST1",
            "GOOGLE_STORAGE_US_EAST2",
            "GOOGLE_STORAGE_US_EAST3",
            "GOOGLE_STORAGE_US_WEST1"
          ]
        },
        "contentType": {"type": "string"},
        "timeModified": {
          "format": "date-time",
          "type": "string"
        },
        "userMetadata": {
          "additionalProperties": {"type": "string"},
          "type": "object"
        }
      }
    },
    "ContainerRequest": {
      "type": "object",
      "properties": {
        "creationMode": {
          "description": "The way a container is creation. If the container should be new, is allowed already exist or use an existing remote container.",
          "type": "string",
          "enum": [
            "NEW_ONLY",
            "REQUIRE_EXISTING",
            "ALLOW_EXISTING"
          ]
        },
        "name": {"type": "string"},
        "backendId": {
          "description": "The backend in which the container will be stored or is stored.",
          "type": "string"
        }
      }
    }
  },
  "swagger": "2.0",
  "info": {
    "license": {
      "name": "Apache License Version 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    },
    "contact": {
      "name": "Sphereon DevOps Team",
      "url": "https://sphereon.com",
      "email": "dev@sphereon.com"
    },
    "description": "Sphereon Storage APIs provide an abstraction layer for storage of containers (folders) and objects (files), using a filesystem backend or AWS-s3, Azure blobstorage or Google cloud storage backend.",
    "termsOfService": "https://sphereon.com",
    "title": "Storage",
    "version": "0.8"
  },
...
}